feat: use anonymous principal by default in HttpAgent (#46)
If the auth transform is missing and the principal is anonymous, use an
anonymous auth transform. Otherwise, use the auth transform as normal.

# BREAKING CHANGE
Before, an auth transform would not be called if it was missing, now it
results as an error if the principal is not anonymous. This was because
just using the packet as is without putting it in an envelope was a bug
and should never be sent to a replica anyway. Now this is explicit.