feat: use anonymous for the first retrieve
Also include some cleanup of the scripts that arent used, and add a diagram for the
bootstrapping process.

This does not affect the worker yet. Only the first retrieve of the index.js on
localhost.

BREAKING CHANGE

If a canister has a retrieve() method that relies on Principal, the principal used
will change.