feat: remove public key and auth transforms and add Identity interface (#67)
The Ed25519 identity implementation has been moved to @dfinity/authentication,
with auth transform to be implemented by Identity. This simplifies creation
and usage of Agents, but is a huge breaking change for code that depends on
@dfinity/agent (although still no breaking changes

An Identity differs from the Public Key interface from before as it is expected
to be able to sign and create principals. It does not have to be able to
expose the secret key blob, but it could (e.g. Ed25519 identity could expose
a BIP-39 export and import functions).

The goal here is to setup the framework to add custom identities, e.g. when
we have COSE implementation, while simplifying agents in general. Also, when
we have delegation, identities should be able to delegate themselves to
other identities, and just work at the higher level of abstraction over
simply key blobs.