feat: implement a WebAuthnIdentity that uses navigator.credentials (#107)
Unfortunately, there is no polyfill existing for navigator.credentials. This has been tested manually against the ic-ref and the replica.

To use the WebAuthn identity in bootstrap, pass in a query argument `_identity=webauthn`. To use a delegated WebAuthn identity, use `_identity=delegate`. To save a WebAuthn key and delegate to a session key for each page load, use `_identity=save`.

This also fixes an error with delegation on real replica. We use the BigInts CBOR extension, but replica requires u64. The spec is not clear on it.