feat: remove logging mentions (#196)
https://github.com/dfinity/agent-js/pull/196#issue-573809771
"""
In the case of the demo those should use console logs. In the general case,
we should consider using an existing library or (only if needed) implement
our own library separately from boostrap.

There was no advantage to adding our own log code. Additionally, there was
a lot of noise in the console, and as a library we shouldnt add much noise
to other peoples code.

The only behaviour changed was that in some cases where removing the
logging left with an empty function, the empty function was removed.
"""