chore: remove scripts and move closer to lerna
This includes a change to request_id from the identity-provider PR
(#132) that I thought was related to something else, but is essential
here with the new lerna packages.

This moves the CI to using lerna entirely, and prevents running
npm install (instead telling the user to run npx lerna bootstrap).