docs: simplify the sample-javascript
Remove authClient and move everything into 3 functions; login, logout and handleCallback.
The goal is to move the login and handleCallback in the authentication package
and be the low level API. Even lower than that, login will most likely be
createRequestUrl() or something to that effect, and leave the navigation to the
application layer.