CI: fix cabal hash
the Github Actions configuration for the Cabal cache was not good
(despite it being the recommended configuration on the cache action
documentation; it’s just broken). The cache key is too fixed, so if
`ic-ref` changes, you still get a cache hit with the old store, and
still rebuild lots of dependendencies, but don’t push a new cache.

This tries to improve this, by including some of cabals configuration
files in the cache key.