fix: use provisional_create_canister_with_cycles and other QoL (#325)
* fix: use provisional_create_canister_with_cycles for creating canisters and other QoL

We should not allow people to run create_canister outside of its
intended context (inter-canister calls). Since this library should
only be used in clients, create_canister should really not be there.

This pushes the release to be more 0.15.6 compatible. Also included
is removing the install_code and create_canister call types, as those
are deprecated since the introduction of the management canister. A
bit overdue for a PR.

* The matrix variable is called `spec`, not `icrefgitref`

Co-authored-by: Joachim Breitner <mail@joachim-breitner.de>