chore: create a new sw-cert app for the service worker (#399)
These are just files that build to a service worker that gets installed and ALWAYS return "Hello World". This is a vague copy of Dom's code, with some comments removed and refactoring in. A lot still need to happen but this should get us started.

The Service Worker right now just does the same as the icx-proxy, but in browser.