refactor: optimize the vec nat8 to copy bytes
Before we ran through the decoding of ALL bytes 1-by-1. With this
change we will read and copy the bytes as buffers. This cuts the
decode and encode performance by A LOT for large files. Since we
use IDL now for making queries to the IC (in the Service Worker),
this makes a huge performance difference on all queries.