fix: Various fixes related to the Service Worker validation (#414)
The new asset canister allows users to have multiple asset encodings, but only one certification (for now). It means that if the asset canister returns a `Content-Encoding: gzip` file the certification might be for the unzipped content. Thus, we need to refactor a bit how content is used in the validation code.

Now the algorithm is basically; decode the body (because we need to return the decoded body anyway to the browser), check that its encoded content is certified (because that's what NNS and Identity do), if not, use the same certificate/tree to check if its identity content is certified.

Also added a new check for `/index.html` if the SHA isn't found. If both aren't found, we invalidate the response (with an error message for debugging).

Also, with the latest agent and using localhost, we cannot use the hard coded mainnet public key, and fetch it. This only works if the SW was installed from `localhost`.