feat(sw): add a compile time env var to always fetch the root key
The environment variable is `FORCE_FETCH_ROOT_KEY` which should be set prior to
building the service worker. It will force the service worker to fetch the root
key on every queries.

An optimization in the HttpAgent has been added to reduce the amount of fetch;
if the root key was fetched once we do not fetch it again. This is safe because
the Agent can only point to a single Host URL, and the key should not change
with the same backend replica.