fix: Remove ArrayBuffer checks from WebAuthnIdentity (#857)
* fix: Remove ArrayBuffer checks from WebAuthnIdentity

This PR removes `instanceof ArrayBuffer` checks from the
`WebAuthnIdentity`. The reason is that these checks hide errors
and cause problem in conjunction with the Bitwarden password
manager (which sets these fields to `Uint8Array`). See also this issue:
https://github.com/dfinity/internet-identity/issues/2235

Simply removing the checks solved the issue entirely, as the fields that
supposedly need to be `ArrayBuffer` are converted to `Uint8Array` anyway.

This change was tested with Internet Identity and the Bitwarden extension
in Chrome.

* bufFromBufLike

* checking presence of attestationObject instead of checking for ArrayBuffer

---------

Co-authored-by: Kyle Peacock <kylpeacock@gmail.com>